make sure array indices are ints

+++ scripts/mrms/mrms_rainrate_comp.py
@@ -49,8 +49,8 @@ def do( now , realtime, delta):
         val = val / 60.0 * 2.0 * 10.0
         val = np.where(val < 0., 255., val)
         ysz, xsz = np.shape(val)
-        x0 = (tilemeta['ul_lon'] - util.WEST) * 100.0
-        y0 = round((tilemeta['ll_lat'] - util.SOUTH) * 100.0,0)
+        x0 = int((tilemeta['ul_lon'] - util.WEST) * 100.0)
+        y0 = int(round((tilemeta['ll_lat'] - util.SOUTH) * 100.0,0))
         imgdata[y0:(y0+ysz),x0:(x0+xsz)] = val.astype('int')
     if count < 4:
         return
